+++ src/Sql/Sql/DataClassification/Model/SensitivityLabelModel.cs
@@ -38,6 +38,9 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Model
         [Ps1Xml(Target = ViewControl.List)]
         public string InformationType { get; set; }
 
+        [Ps1Xml(Target = ViewControl.List)]
+        public SensitivityRank? Rank { get; set; }
+
         [Hidden]
         public string SensitivityLabelId { get; set; }
 
@@ -46,8 +49,7 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Model
 
         public override string ToString()
         {
-            StringBuilder builder = new StringBuilder();
-            builder.AppendLine("{");
+            List<string> valuesPerPropertyName = new List<string>();
             foreach (var property in this.GetType().GetProperties())
             {
                 string name = property.Name;
@@ -59,10 +61,13 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Model
                 object value = property.GetValue(this);
                 if (value != null)
                 {
-                    builder.AppendLine($"\t{name}: {value},");
+                    valuesPerPropertyName.Add($"\t{name}: {value}");
                 }
             }
 
+            StringBuilder builder = new StringBuilder();
+            builder.AppendLine("{");
+            builder.AppendLine(string.Join($",{Environment.NewLine}", valuesPerPropertyName));
             builder.Append("}");
 
             return builder.ToString();
@@ -96,25 +101,33 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Model
                 }
                 else
                 {
-                    throw new Exception($"Information Type '{newInformationType}' is not part of Information Protection Policy. Please add '{newInformationType}' to the Information Protection Policy, or use one of the following: {ToString(informationProtectionPolicy.SensitivityLabels.Keys)}");
+                    throw new Exception($"Information Type '{newInformationType}' is not part of Information Protection Policy. Please add '{newInformationType}' to the Information Protection Policy, or use one of the following: {ToString(informationProtectionPolicy.InformationTypes.Keys)}");
+                }
+            }
+        }
+        private void ApplySensitivityLabel(string newSensitivityLabel, InformationProtectionPolicy informationProtectionPolicy)
+        {
+            if (!string.IsNullOrEmpty(newSensitivityLabel) &&
+                !string.Equals(SensitivityLabel, newSensitivityLabel))
+            {
+                if (informationProtectionPolicy.SensitivityLabels.TryGetValue(newSensitivityLabel, out Tuple<Guid, SensitivityRank> idRankTuple))
+                {
+                    SensitivityLabel = newSensitivityLabel;
+                    SensitivityLabelId = idRankTuple.Item1.ToString();
+                    Rank = idRankTuple.Item2;
+                }
+                else
+                {
+                    throw new Exception($"Sensitivity Label '{newSensitivityLabel}' is not part of Information Protection Policy. Please add '{newSensitivityLabel}' to the Information Protection Policy, or use one of the following: {ToString(informationProtectionPolicy.SensitivityLabels.Keys)}");
                 }
             }
         }
 
-        private void ApplySensitivityLabel(string newSensitivityLabel, InformationProtectionPolicy informationProtectionPolicy)
+        private void ApplySensitivityRank(SensitivityRank? newSensitivityRank)
         {
-            if (!string.IsNullOrEmpty(newSensitivityLabel) ||
-                !string.Equals(SensitivityLabel, newSensitivityLabel))
+            if (newSensitivityRank != null && newSensitivityRank != Rank)
             {
-                if (informationProtectionPolicy.SensitivityLabels.TryGetValue(newSensitivityLabel, out Guid sensitivityLabelId))
-                {
-                    SensitivityLabel = newSensitivityLabel;
-                    SensitivityLabelId = sensitivityLabelId.ToString();
-                }
-                else
-                {
-                    throw new Exception($"Sensitivity Label '{newSensitivityLabel}' is not part of Information Protection Policy. Please add '{newSensitivityLabel}' to the Information Protection Policy, or use one of the following: {ToString(informationProtectionPolicy.InformationTypes.Keys)}");
-                }
+                Rank = newSensitivityRank;
             }
         }
 
@@ -122,5 +135,14 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Model
         {
             return string.Join(", ", collection.Select(s => $"'{s}'"));
         }
+    }
+
+    public enum SensitivityRank
+    {
+        None,
+        Low,
+        Medium,
+        High,
+        Critical
     }
 }